Implementation of Add-Appointment API

+++ src/main/java/edu/icet/crm/service/AppointmentService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface AppointmentService {
     List<Appointment> getAll();
     void addAppointment(Appointment appointment);
-    void deleteAppointmentById(Integer id);
-    Appointment searchAppointmentById(Integer id);
+    void deleteAppointmentById(Long id);
+    Appointment searchAppointmentById(Long id);
     void updateAppointmentById(Appointment appointment);
 }
